Support Noise (Other Channel)
Added to episodes 4+5 only for now, as these are the affected episodes. Will add to the remaining for next week if this is working well.

NOC code is adapted from WUSC-KEEP-II. The main significant change has been adapting it for a pipeline which supports running in auto-code mode only, as WUSC doesn't support this. On WUSC the strategy is to generate the production file after applying the manual codes. Since that doesn't work for an auto-code/full-run set-up, this is currently set to produce the production file before running the noise filter, but happy to discuss.