Add new noise codes to coda datasets
These let us label by the type of 'noise' we're seeing - either people trying to get in touch with a pregnancy support program or seeking coronavirus cash transfer funds. We label these separately so that we have the option to explain to people they reached the wrong number if needed. The existing Noise (Other Channel) code is maintained in case we see any other cases that don't fall into those new categories.

Note that this update is to code schemes only for now. Will update the pipeline to process the new noise categories in a future PR